attendance_app/requirements.txt
flask
requests
qrcode
pillow
pandas
openpyxl
pyinstaller
pyzbar